demo/unstable.html

<!doctype html>
<title>Unstable</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  assert_true(Math.random() > 0.5);
}, 'known unstable');
</script>
